requirements.txt
# Flask and extensions
Flask==3.1.1
flask-sqlalchemy==3.1.1
flask-migrate==4.0.6
flask-login==0.6.3
flask-wtf==1.2.1
flask-cors==4.0.0
email-validator==2.1.1

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.27
alembic==1.12.1

# Testing
pytest==8.3.5
pytest-flask==1.3.0
coverage==7.8.1

# Deployment
gunicorn==21.2.0
python-dotenv==1.0.1
pyngrok==7.2.0

# Utilities
werkzeug==3.1.3
requests==2.31.0
itsdangerous==2.2.0
Jinja2==3.1.6
markdown==3.5.2
python-dateutil==2.9.0
pytz==2024.1
pyjwt==2.8.0
cryptography==42.0.5

# Logging
python-json-logger==2.0.7

# Development
black==24.4.2
flake8==7.2.0
isort==5.13.2
